# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "一个简单的记账系统 API，支持用户注册、登录、消费记录管理和数据导出功能",
        "title": "记账系统 API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/admin/ai-analysis": {
            "post": {
                "description": "选择时间范围和AI模型，对消费记录进行AI分析，SSE流式返回JSON帧（delta/done/error）。管理员可分析所有记录或指定用户的记录（通过user_id参数），非管理员只能分析自己的记录。分析结束后会保存到历史记录。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "后台管理-AI分析"
                ],
                "summary": "AI分析消费记录（流式）",
                "parameters": [
                    {
                        "description": "分析请求（user_id字段仅管理员可用）",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.AnalysisRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SSE流：data: {\\\"type\\\":\\\"delta\\\",\\\"content\\\":\\\"...\\\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "参数错误或该时间范围内没有消费记录",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "AI模型不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/admin/ai-analysis/history": {
            "get": {
                "description": "获取AI分析历史记录，按model_id分页返回（软删除不返回）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台管理-AI分析"
                ],
                "summary": "获取AI分析历史",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "AI模型ID",
                        "name": "model_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "页码，默认1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页条数，默认20，最大100",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功，返回分页数据",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/admin/ai-analysis/history/{id}": {
            "delete": {
                "description": "软删除指定的AI分析历史记录",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台管理-AI分析"
                ],
                "summary": "删除AI分析历史",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "历史记录ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "无效的ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "记录不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/admin/ai-chat": {
            "post": {
                "description": "选择AI模型，与AI进行对话，SSE流式返回JSON帧（delta/done/error）。结束后保存聊天记录。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "后台管理-AI聊天"
                ],
                "summary": "AI聊天（流式）",
                "parameters": [
                    {
                        "description": "聊天请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.AIChatRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SSE流：data: {\\\"type\\\":\\\"delta\\\",\\\"content\\\":\\\"...\\\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "AI模型不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/admin/ai-chat/history": {
            "get": {
                "description": "获取AI聊天历史记录，按model_id分页返回（软删除不返回）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台管理-AI聊天"
                ],
                "summary": "获取AI聊天历史",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "AI模型ID",
                        "name": "model_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "页码，默认1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页条数，默认20，最大100",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功，返回分页数据",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/admin/ai-chat/history/{id}": {
            "delete": {
                "description": "软删除指定的AI聊天记录",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台管理-AI聊天"
                ],
                "summary": "删除AI聊天记录",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "聊天记录ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "无效的ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "记录不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/admin/ai-models": {
            "get": {
                "description": "获取系统中所有AI模型配置列表（不包含APIKey）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台管理-AI模型"
                ],
                "summary": "获取AI模型列表",
                "responses": {
                    "200": {
                        "description": "获取成功，返回模型列表",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "创建新的AI模型配置，包括名称、API地址和密钥",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台管理-AI模型"
                ],
                "summary": "创建AI模型",
                "parameters": [
                    {
                        "description": "AI模型信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CreateAIModelRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "参数错误或模型名称已存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/admin/ai-models/reorder": {
            "put": {
                "description": "根据传入的模型ID顺序更新排序，用于前端拖拽排序后保存",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台管理-AI模型"
                ],
                "summary": "排序AI模型",
                "parameters": [
                    {
                        "description": "模型ID顺序",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ReorderAIModelsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "排序成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/admin/ai-models/{id}": {
            "get": {
                "description": "根据ID获取AI模型配置详情（不包含APIKey）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台管理-AI模型"
                ],
                "summary": "获取单个AI模型",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "AI模型ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "无效的ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "模型不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "description": "更新指定的AI模型配置信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台管理-AI模型"
                ],
                "summary": "更新AI模型",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "AI模型ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新的模型信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.UpdateAIModelRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "参数错误或模型名称已存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "模型不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "description": "删除指定的AI模型配置（软删除）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台管理-AI模型"
                ],
                "summary": "删除AI模型",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "AI模型ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "无效的ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "模型不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/admin/ai-models/{id}/test": {
            "post": {
                "description": "向AI模型发送轻量测试请求，检测接口是否可用",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台管理-AI模型"
                ],
                "summary": "检测AI接口可用性",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "AI模型ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "检测成功，接口可用",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "无效的ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "模型不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "502": {
                        "description": "接口不可用",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/admin/categories": {
            "get": {
                "description": "获取所有消费类别列表，支持按名称模糊搜索",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台管理-消费类别"
                ],
                "summary": "获取消费类别列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "类别名称（模糊匹配）",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功，返回类别列表",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "创建新的消费类别，支持设置名称、排序和颜色",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台管理-消费类别"
                ],
                "summary": "创建消费类别",
                "parameters": [
                    {
                        "description": "类别信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CategoryCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "参数错误或类别名称已存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/admin/current-user": {
            "get": {
                "description": "获取当前登录用户的详细信息（包括用户ID）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台管理"
                ],
                "summary": "获取当前登录用户信息",
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/admin/export/excel": {
            "get": {
                "description": "根据时间范围导出消费记录为Excel文件。管理员可导出所有用户数据，普通用户只能导出自己的数据。",
                "produces": [
                    "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
                ],
                "tags": [
                    "后台管理-导出"
                ],
                "summary": "导出消费记录为Excel",
                "parameters": [
                    {
                        "type": "string",
                        "description": "开始时间 (YYYY-MM-DD)",
                        "name": "start_time",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "结束时间 (YYYY-MM-DD)",
                        "name": "end_time",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Excel文件",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.AIChatRequest": {
            "type": "object",
            "required": [
                "message",
                "model_id"
            ],
            "properties": {
                "message": {
                    "type": "string",
                    "minLength": 1
                },
                "model_id": {
                    "type": "integer"
                }
            }
        },
...
        "api.AnalysisRequest": {
            "type": "object",
            "required": [
                "end_time",
                "model_id",
                "start_time"
            ],
            "properties": {
                "end_time": {
                    "type": "string",
                    "example": "2024-12-31"
                },
                "model_id": {
                    "type": "integer"
                },
                "start_time": {
                    "type": "string",
                    "example": "2024-01-01"
                },
                "user_id": {
                    "description": "可选，仅管理员可用，用于筛选指定用户的账单",
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "api.CreateAIModelRequest": {
            "type": "object",
            "required": [
                "api_key",
                "base_url",
                "name"
            ],
            "properties": {
                "api_key": {
                    "type": "string",
                    "minLength": 1,
                    "example": "sk-..."
                },
                "base_url": {
                    "type": "string",
                    "example": "https://api.openai.com/v1"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1,
                    "example": "OpenAI GPT-4"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}